Use close() instead of closesocket() to prevent FD leaks on windows.
This is due to ruby.h/defines.h/win32.h in the ruby source redefining
all the socket functions to call rb_w32_* functions instead. This means
the return value of socket() is not a win32 SOCKET handle, but instead
a file descriptor wrapper that the ruby VM owns. We must call close()
(which invokes rb_w32_close) to dispose of it correctly.