EventMachine_t::TimerBindings is a map taking timer bindings to firing times.
It is inserted into and removed from whenever Timers is, effectively becoming
a reversed map.

It is used in the new method EventMachine_t::UninstallOneshotTimer, which
accepts a timer signature (binding) and removes the corresponding entry from
Timers and TimerBindings (returning true if something is removed.)

TimerBindings enables UninstallOneshotTimer to avoid a linear scan of Timers,
instead making it log(Timers.size()).