Added EventMachine::Reactor#uninstall_oneshot_timer for pure-ruby.
It uses a hash instead of a (tree) map for @timer_bindings, but otherwise
should be equivalent to the prior c++ changes.

This change is necessary since, currently, timer cancelation is handled by
EM.cancel_timer in front of both pure ruby and c++ impls.  If that
responsibility shifts to c++, then pure-ruby must also implement it.