Delete Java reactor's associated java tests and application layer. These are out of date, unmaintained and there is no point in keeping them around.
The Java reactor (and the C++ reactor) were both initially designed to
be usable outside of Ruby. However, no one actually does and keeping (or
pretending to keep) compatibility layers just slows down development and
makes the experience of using and developing EM for ruby more painful. I
would prefer instead to focus development efforts on the Ruby use-case
specifically.

That said, these files are and will remain in the Git history if someone
wishes to use them.