Remove unsupported C++ API.
This is a C++ API layer built on top of the evma_* C API layer (which
itself is built upon the core C++ reactor API). This code is
unmaintained and unused (like the Java API in
d8fa9897d1d468e676513c74840e8640fc1fc28d), so I am removing it.