Remove unused win32 version of em.cpp
This appears to be a work-in-progress copy of em.cpp that was supposed
to add support for IoCompletionPort on windows.

If someone wishes to implement an IOCP mode for EM in the future, they might
benefit from this file. However, I would recommend instead that IOCP be added
as a mode to the original reactor (similar to the way epoll and kqueue
work) with ifdefs where necessary, such that a complete separate copy of
the reactor does not have to be maintained for win32.