EM::Connection#unbind takes an optional error reason argument.
The reason argument is an Errno::E* class, or nil when the error was
unknown (usually due to a clean shutdown). I only tested the TCP
codepaths, so the UDP and unix socket code paths will not propagate
errors correctly in all cases.

EM::Connection#error? now also returns an Errno::E* class instead of
true.

Finally, I noticed that errors on read() were not correctly being
handled, making it possible for errors like ECONNRESET and ENOTCONN to
be ignored. In these cases, EM would never fire an unbind on the
connection.