Add a basic dynamic environments implementation
If an exception is raised in the middle of a callback (and not handled
in the callback), the only opportunity the programmer has to deal with
it is in the global error handler. However, in the error handler, the
context in which the error was raised has been lost. With this patch,
callbacks can store values in dynamic environments. If an exception is
raised, the global error handler can access the same dynamic
environment.

USAGE:

'EventMachine::Environment.dynamic_env' returns the current dynamic
env, which can be accessed like a hash.

EXAMPLE:

<<EOF
require 'rubygems'
require 'eventmachine'

EventMachine.error_handler do |er|
  puts "Foo key is #{EventMachine::Environment.dynamic_env[:foo].inspect}"
  EventMachine.stop
end

EventMachine.next_tick do
  EventMachine::Environment.dynamic_env[:foo] = :bar
  raise 'Raised!'
end

EventMachine.run
EOF

CAVEATS:

- This hasn't been tested extensively. I'm sure it doesn't work in
  corner cases. I wouldn't be shocked if it doesn't work in most
  normal cases.

- The code is pretty bad right now. I'll fix it up later once it's
  closer to a final version.