ssl_verify_wrapper now checks CN against expected hostname
This hostname is passed in as a :hostname option to start_tls.
It's stored on the SslBox_t, using ex_data.

I'm not 100% using sk_X509_shift is the best way to go about
getting the correct certificate to check against, but it seems to work
in my test cases.

I will shortly rewrite these test cases to not hit the network.