Clean up the threadpool on forking a new reactor
When you fork a new reactor instance, the old eventmachine instance
gets destroyed and recreated (in the c) with release_machine. However,
because the threadpool is held inside the ruby code, the new machine
has the (dead) threads associated with the old instance, which can
fill up the threadpool and cause the new reactor to not be able to
execute any defer calls.