Ruby 2.0 may fail at Queue.pop, so rescue and complain to $stderr
The message is "Attempt to unlock a mutex which is not locked (ThreadError)"
and it is caused by a bug in Ruby 2.0. Annoying people with $stderr is arguably
better than crashing their app. That said, it might not be worth the loud error
because the Queue does not appear to drop any tasks even when this occurs.