Connection::error? calls report_connection_error_status in Ruby reactor (#801)
This method exists in the cpp version but not in the pure ruby version. When
there is an undelivered message for example, it will attempt to check the socket
status and fail saying that the method report_connection_error_status does not
exist. This is in turn causes the connection to fail. It will not return the
socket status.