Switch from epoll_create to epoll_create1
The original epoll_create took a size argument as a hint for how much
kernel space to allocate for events. Since Linux 2.6.8, the size
argument is ignored, but must be greater than zero.

Since we also need to set CLOEXEC on the epoll fd, there is a small
window between epoll_create and fcntl as they are separate function
calls. Since Linux 2.6.27 and glibc 2.9, epoll_create1 accepts the
CLOEXEC argument at creation time to eliminate this gap.

This drops support for epoll on RHEL 5 / CentOS 5, EOL was March 2017.

Resolves #770