The unbind callback must not raise exceptions
The unbind callback absolutely must not raise an exception or the reactor will crash.
If there is no EM.error_handler, or if the error_handler retrows, then stop the reactor,
stash the exception in $wrapped_exception, and the exception will be raised after the
reactor is cleaned up (see the last line of self.run).

This fixes many crash reports in the 1.2 series since #327 was merged.