Fix crashes on Max OS X on large numbers of connections
All three fd_set arguments to select must be large enough to modify
according to the length given in the first argument to select. This can
be done by calling rb_fd_resize on each of the fd_set arguments, which
happens to be exactly what rb_fd_select does!