Fix pure ruby ssl_verify_peer callback
The pure ruby version was only calling SslVerify once, after
SslHandshakeCompleted, which is out of order and won't allow the
connection to see the entire certificate chain.

This adds the SSLContext#verify_callback to trigger the
`ssl_verify_peer` method, to provide feature parity with the cpp
implementation.  To properly handle the callback, we need to pass the
correct arguments, but that's out of scope for this (a fix for that is
in a different PR).