Raise CI timeout intervals for windows and macos
Several of these tests appeared to be flaky in CI, but only for macos or
windows.  So I went through and raised many of them, for those OSes
only.  In many cases, I changed the timeouts to be a multiple of
TIMEOUT_INTERVAL, so that individual tests don't each need to check if
they are running on `windows? || darwin?`.