Switch CI to use macos-latest runner
GitHub Actions dropped support for macos-10.15 on 2022-12-01:
- https://github.com/actions/runner-images/issues/5583

Under the presumption that MacOS is primarily used for development and
not for deployment, I've switched the macos GitHub Actions runners to
use the "latest" release version for ruby >= 2.6.  Earlier ruby's don't
support arm64, so they are using the macos-13 runner (which is amd64).

On the one hand, this is less explicit and is no longer "pinned", so it
may result in CI failing when GitHub Actions updates these runners.  On
the other hand, we also won't need another PR to update these every time
an old version is dropped or a new version is added.

I attempted to make the same change from windows-2019 to windows-latest.
But I saw three tests that reliably pass under windows-2019 and reliably
fail under windows-2022.  Since I don't have a Windows OS dev env at my
disposal, I'm opting to just let windows CI continue under windows-2019,
for now.