WIP: Attempt (and fail) to fix macos text_connrefused
This converts the conditional for platform specific code from a regexp
on RUBY_PLATFORM to instead check for all of the relevant consts.  I
thought that this would allow us to see the tcpi_state for both Linux
and MacOS... but on the MacOS machines I tested,
Socket::TCP_CONNECTION_INFO was missing.