Add pure ruby stub methods (raise EM::Unsupported)
Many of these may be trivial to implement in pure ruby mode, and some
may be impossible, but for right now I'd prefer to simply have them all
displayed as a "TODO list" in the test results.

_This would have broken many tests_ that were previously omitted based
on whether or not `EM` responds to these methods.  However an earlier
commit marked those tests as pending in pure_ruby mode.