Add EM.current_time for pure_ruby and enable tests
This already existed on Reactor#current_loop_time, we just needed to
forward to it from the singleton method on `EM`.  For now, I kept the
same name for the Reactor attr_reader but added an alias to it.

This enables four existing tests.  They are very slow in pure ruby
mode, but they do pass!