Run pure_ruby tests in same job as regular tests
It feels the builds get randomly hit by one or more job that just go a
bit slower than the others which impacts the flakier tests.  With so
many jobs in the matrix, builds are almost guaranteed to get hit by at
least one of them behaving so badly that they fail.  Perhaps, by merging
the `pure_ruby` step into the build job, the odds of every job in the
matrix completing on the first attempt are improved.